Fix potential overflow in getSizeInBytes

+++ presto-spi/src/main/java/com/facebook/presto/spi/block/ByteArrayBlockBuilder.java
@@ -125,11 +125,10 @@ public class ByteArrayBlockBuilder
         }
     }
 
-    // Copied from ByteArrayBlock
     @Override
     public long getSizeInBytes()
     {
-        return (Byte.BYTES + Byte.BYTES) * positionCount;
+        return (Byte.BYTES + Byte.BYTES) * (long) positionCount;
     }
 
     @Override
